# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API Centro de Costos - Cámara de Comercio
  description: |
    API REST para la gestión de registro documental, facturas, órdenes y procesos contables.
    
    ## Arquitectura
    Esta API está construida con Node.js, Express y Sequelize, siguiendo el patrón de arquitectura Repository.
    
    ## Autenticación
    Por definir en futuras versiones.
    
    ## Base de datos
    PostgreSQL - Solo operaciones DML permitidas (SELECT, INSERT, UPDATE, DELETE)
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: desarrollo@example.com
  license:
    name: ISC

servers:
  - description: Servidor Local de Desarrollo
  - description: Servidor de Producción

tags:
  - name: Health
    description: Endpoints de verificación del estado del servidor
  - name: Facturas
    description: Operaciones relacionadas con facturas y registro documental

paths:
  /health:
    get:
      tags:
        - Health
      summary: Verificar estado del servidor
      description: Endpoint para verificar que el servidor esté funcionando correctamente
      operationId: getHealth
      responses:
        '200':
          description: Servidor funcionando correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: Servidor funcionando correctamente
                  timestamp:
                    type: string
                    format: date-time
                    example: 2025-10-29T18:30:00.000Z

  /api/invoices/documentary-record-single-order:
    get:
      tags:
        - Facturas
      summary: Obtener registro documental de orden única
      description: |
        Obtiene todas las facturas que cumplen con los siguientes criterios:
        - **Dependencia**: Debe ser 'Provedor' o 'Cuenta Cobro'
        - **Estado**: Debe estar 'Radicada'
        - **Orden**: No debe estar 'Vacio'
        - **Multiorden**: Debe ser 'Pendiente' o 'No'
        
        Este endpoint es útil para obtener facturas que requieren procesamiento de orden única.
      operationId: getDocumentaryRecordSingleOrder
      responses:
        '200':
          description: Registro documental obtenido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Registro documental de orden única obtenido correctamente
                  count:
                    type: integer
                    description: Número total de registros encontrados
                    example: 0
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/invoices/documentary-record-multi-order:
    get:
      tags:
        - Facturas
      summary: Obtener registro documental de multiorden
      description: |
        Obtiene todas las facturas que cumplen con los siguientes criterios:
        - **Dependencia**: Debe ser 'Provedor' o 'Cuenta Cobro'
        - **Estado**: Debe estar 'RadicadaMultiorden'
        - **Multiorden**: Debe ser 'Si'
        
        Este endpoint es útil para obtener facturas que requieren procesamiento de multiorden.
      operationId: getDocumentaryRecordMultiOrder
      responses:
        '200':
          description: Registro documental de multiorden obtenido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Registro documental de multiorden obtenido correctamente
                  count:
                    type: integer
                    description: Número total de registros encontrados
                    example: 0
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/invoices/documentary-record:
    get:
      tags:
        - Facturas
      summary: Obtener registro documental completo (orden única y multiorden)
      description: |
        Obtiene todas las facturas separadas por tipo para facilitar el procesamiento por RPA.
        
        **Retorna dos conjuntos de datos:**
        - **singleOrder**: Facturas con orden única (estado='Radicada', multiorden IN 'Pendiente'/'No')
        - **multiOrder**: Facturas con multiorden (estado='RadicadaMultiorden', multiorden='Si')
        
        **IMPORTANTE:** Cada factura retorna TODOS los campos del modelo Invoice completo (id, numeroFactura, nit, orden, nombre, importe, iva, dependencia, estado, procesado, multiorden, fondo, radicado, subtotal, fechaValidacion, descripcion, etc.).
        Los ejemplos mostrados están simplificados por legibilidad, pero la respuesta real incluye aproximadamente 40 campos por factura.
        
        Este endpoint combina ambas consultas en una sola petición, optimizando el procesamiento para sistemas RPA.
      operationId: getDocumentaryRecordAll
      responses:
        '200':
          description: Registro documental completo obtenido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Registro documental completo obtenido correctamente
                  total:
                    type: integer
                    description: Número total de facturas (orden única + multiorden)
                    example: 0
                  singleOrder:
                    type: object
                    properties:
                      count:
                        type: integer
                        description: Número de facturas con orden única
                        example: 0
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'
                  multiOrder:
                    type: object
                    properties:
                      count:
                        type: integer
                        description: Número de facturas con multiorden
                        example: 0
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'

        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/invoices/documentary-record/send-to-rpa:
    post:
      tags:
        - Facturas
      summary: 🤖 TEST - Enviar facturas al RPA
      description: |
        **Endpoint de prueba** para enviar facturas al sistema RPA (FastAPI) para procesamiento automático.
        
        **Flujo:**
        1. Obtiene todas las facturas (orden única y multiorden)
        2. Envía los datos al sistema RPA en FastAPI
        3. Retorna confirmación del envío
        
        **⚠️ Prerrequisitos:**
        - La API FastAPI (JSP7) debe estar corriendo en `http://localhost:8000`
        - El endpoint `/api/v1/sequences/registro_documental/execute` debe estar disponible
        
        **🔄 Uso:** Este endpoint es para testing de la integración entre ambas APIs.
      operationId: sendToRPA
      responses:
        '200':
          description: Datos enviados al RPA correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Proceso completado
                  invoices:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 0
                      singleOrderCount:
                        type: integer
                        example: 0
                      multiOrderCount:
                        type: integer
                        example: 0
                  rpa:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: Datos enviados al RPA correctamente
                      rpaResponse:
                        type: object
        '500':
          description: Error al procesar la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/invoices/rpa/status:
    get:
      tags:
        - Facturas
      summary: 🔍 Verificar estado del RPA
      description: |
        Verifica si el servicio RPA (FastAPI) está disponible y respondiendo.
        
        **Uso:** Ejecuta este endpoint antes de enviar datos para asegurar conectividad.
        
        **URL FastAPI:** Configurada en variable de entorno `RPA_API_URL`
      operationId: checkRPAStatus
      responses:
        '200':
          description: Servicio RPA disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Servicio RPA disponible
                  rpaStatus:
                    type: object
        '503':
          description: Servicio RPA no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Servicio RPA no disponible

components:
  schemas:
    Invoice:
      type: object
      description: Modelo de factura con toda la información de facturación
      properties:
        id:
          type: integer
          description: ID único de la factura
        numeroFactura:
          type: string
          description: Número de la factura
        nit:
          type: string
          description: NIT del proveedor
        orden:
          type: string
          description: Número de orden asociada
        tipoOrden:
          type: string
          description: Tipo de orden (Compra, Servicio, etc.)
        nombre:
          type: string
          description: Nombre del proveedor
        importe:
          type: integer
          description: Importe total de la factura
        iva:
          type: number
          format: double
          description: Valor del IVA
        dependencia:
          type: string
          description: Dependencia que procesa la factura
        estado:
          type: string
          description: Estado actual de la factura
        procesado:
          type: string
          description: Estado de procesamiento
        fechaValidacion:
          type: string
          format: date
          description: Fecha de validación de la factura
        fechaXml:
          type: string
          format: date
          description: Fecha del archivo XML
        pdfBase64:
          type: string
          description: PDF de la factura codificado en base64
        nombreArchivo:
          type: string
          description: Nombre del archivo original
        prefijoFactura:
          type: string
          description: Prefijo de la factura
        subtotal:
          type: number
          format: decimal
          description: Subtotal de la factura
        rechazo:
          type: string
          description: Motivo de rechazo si aplica
        metodoPago:
          type: string
          description: Método de pago
        radicado:
          type: string
          description: Número de radicado
        notaCredito:
          type: string
          description: Indica si tiene nota de crédito
        fondo:
          type: string
          description: Fondo presupuestal
        numero:
          type: string
          description: Número interno
        ano:
          type: string
          description: Año de la factura
        numeroOrden:
          type: string
          description: Número de orden
        multiorden:
          type: string
          description: Indica si es multiorden
        recibo:
          type: string
          description: Número de recibo
        fechaContabilizacion:
          type: string
          format: date
          description: Fecha de contabilización
        descripcion:
          type: string
          description: Descripción de la factura
        numeroCxp:
          type: string
          description: Número de cuentas por pagar
        contrato:
          type: string
          description: Número de contrato asociado
        numeroContabilizado:
          type: string
          description: Número de contabilización
        reteiva:
          type: string
          description: Aplica retención de IVA
        retefuente:
          type: string
          description: Aplica retención en la fuente
        codigoImpuesto:
          type: string
          description: Código del impuesto aplicable
        estadoCamara:
          type: string
          description: Estado en cámara de comercio
        honorariosImpuesto:
          type: string
          description: Impuesto sobre honorarios
        registroDrive:
          type: string
          description: Registro en Google Drive
        eventos:
          type: string
          description: Eventos relacionados
        envioDocumentos:
          type: string
          description: Estado de envío de documentos

    Error:
      type: object
      description: Modelo de respuesta de error
      properties:
        status:
          type: string
        message:
          type: string
        details:
          type: object
          description: Detalles adicionales del error (solo en desarrollo)